Rework widget logic, add functionality of input parsing and restructuring.
WidgetSet class was reworked to take in input in pairings. This would be a pill name, and quantity. This is to allow for displaying of said pairing onto the screen. A total of 8 pairings are allowed to be taken in as a "prescription" and is then parsed to pass into the class. Upon creating the widgets, the parsed prescription is then re-parsed to be sent back out.